image-builder: correct method to obtain the last chunk size
If i_size is exactly multiple times of CHUNK_SIZE. We just to
use file size to minus total size of other chunks. Otherwise,
a chunk is leaked.

In addtion, this patch re-define DEFAULT_RAFS_BLOCK_SIZE as `u64`, so
we don't have to convert it everywhere.

+++ rafs/src/layout.rs
@@ -28,7 +28,7 @@ pub const RAFS_SUPERBLOCK_SIZE: usize = 8192;
 pub const RAFS_INODE_INFO_SIZE: usize = 512;
 pub const RAFS_CHUNK_INFO_SIZE: usize = 128;
 
-pub const DEFAULT_RAFS_BLOCK_SIZE: usize = 1024 * 1024;
+pub const DEFAULT_RAFS_BLOCK_SIZE: u64 = 1024 * 1024;
 pub const RAFS_SUPER_MAGIC: u32 = 0x5241_4653;
 
 const RAFS_XATTR_ALIGNMENT: usize = 8;
